Fixed some bugs. The second pulse counter is now working correctly.The indicator displays the value correctly when switching to the setting mode.

+++ ctrl_timer.h
@@ -35,6 +35,6 @@ void start_timer0();
 void stop_timer0();
 //void add_timer_task(type_queue_node task);
 void init_test_timer_queue(void);
-void add_new_task_with_delay(const TPTR task, const size_t delay, const size_t tiks);
+void add_new_task_with_delay(const TPTR task, const uint16_t delay, const uint16_t tiks);
 
 #endif // CTRL_TIMER_H
